[Event Hubs] Load balancing refactor (#40986)
* [Event Hubs] Load balancing refactor

The focus of these changes is to refactor the load balancing logic to defer
awaiting partitions when ownership is lost until either the processor is
stopping or that same partition has been reclaimed.  This reduces the number
of operations that need to be performed and allows load balancing to continue
running on a predictable schedule.

Previously, when event processing was long-running and did not honor the
cancellation token when ownership was lost, load balancing would be delayed
until the processing completed.  This could result in a pause in ownership
renewals which could lead to partitions being stolen by another processor.

This also allowed removal of existing LINQ expressions and a more efficient
approach to reconciling owned partitions.