[AppConfig] Auth Policy Fix
The focus of these changes is to adjust the authorization policy to make
use of the Authority when computing the auth signature, as this is what is
represented in the host header used to validate.

This works by coincidence when the standard HTTP ports are used, as they
do not appear in the host header, but becomes an issue when a non-standard
port is needed.