[Messaging] Disable Mooncake for weekly test runs
The focus of these changes is to disable the Mooncake (China) region for
weekly test runs, as the tests have been consistently timing out for the
last 6 months.  This is believed to be due to the latency when running in
our DevOps environment and targeting Azure resources in Mooncake.  Until
we have the ability to run within the same cloud being tested, there's no
value in having the cloud enabled.