[Event Hubs] Sample and Doc updates
The focus of these changes is to make formatting changes to the new
sample for using the processor with an ASP.NET hosted service, to better
align with the Event Hubs conventions.

Also included are updates to the docs to point links at "learn.microsoft.com"
and to correct the template deployment link such that it now directly targets
the `main` branch rather than requiring a redirect.

Also included is a correction to a typo in an ETW id that caused a duplicate
registration.