[Messaging] Loosen constructor validation
The focus of these changes is to loosen validation
for the fully qualified namespace name passed to
client constructors.  A URI is now also accepted
as a valid format.  This is intended to improve the
experience when using the management library, CLI,
Bicep, or ARM template to create the namespace, as
they return only an endpoint for the namespace.
Previously, callers were responsible for parsing the
endpoint and extracting the host name for use with
the clients.