Add missing token credentials for ShareServiceClient when configured using dependency injection (#39755)
* Pass token credential to share client builder

When using the Azure client builder `AddFileServiceClient` extension
method to configure a `ShareServiceClient` using the file service URI,
the `TokenCredential` instance was not being passed to the
`ShareServiceClient` constructor. Consequently, instances of
`ShareClient` created from it were unable to authenticate because they
did not have access to the credential configured in the client builder.

This commit ensures that `ShareServiceClient` is constructed with the
credential so that its `ShareClient` instances can use the token for
authentication.

* Address pull request comments

This commit adds a new extension method to the Azure Files Shares client
factory builder that allows for the creation of a `ShareServiceClient`
by URI that uses the configured token credential. This revision avoids
a breaking change that would have been caused by changing the generic
constraint of the existing extention method.

* Disable editor browsing of legacy extension method

This commit annotates the legacy AddFileServiceClient extension method
with `EditorBrowsableState.Never` and adds a comment to the method
explaining that it creates a client that ignores the configured
`TokenCredential` and lacks autehntication credentials.

* Revise comments to remove unnecessary `using`

This commit revises the XML doc comments to remove a reference that is
not strictly needed and may have caused build issues.