[Event Hubs] Change trigger checkpointing behavior (#42299)
* [Event Hubs] Change trigger checkpointing behavior

The focus of these changes is to change the behavior
of the Event Hubs trigger to not checkpoint when the
listener is shutting down. This is necessary to
prevent potential data loss from occurring when
shutting down Function retries.  Because the trigger
cannot know if the Function host would have retried
a failure if it were not shutting down, we cannot
assume that it is safe to checkpoint.  This change
ensures that the batch of events being processed
when shut down was requested will be retried by
another instance or the next time the Function app
is run.  Though this may cause duplicate processing,
it is consistent with the Event Hubs at-least-once
guarantee.