[Event Hubs] Changelogs and trace member rename
The focus of these changes is to add change log
entries for the recent event processor batch tracing
fix and to rename the new member to better align with
existing names in the package.  This also changes
the form of the member slightly, as it does not
need to virtual to be set by derrived processors,
nor does a null value have meaning.