[Template] Remove unused perf and stress folders (#43070)
* [Template] Remove unused perf and stress folders

The focus of these changes is to remove the unused
perf and stress folders from the template.  The
Azure SDK team no long requires performance tests
for each library and partner teams do not use the
scaffolding when generating a new library.  Likewise,
the stress folder is empty, save a README, which
simply points to a wiki location that our onboarding
guide also references.

In the interest of streamlining the template to avoid
confusion and to keep the size of the repository down,
these folders are being removed.