[Messaging] Allow binary application properties (#43181)
* [Messaging] Allow binary application properties

The focus of these changes is to update AMQP conversion
to remove the validation that disallows `byte[]` and
`ArraySegment<byte>` values in application properties.
These are confirmed to be encoded as `binary` by
`Microsoft.Azure.Amqp`, making them allowable values
according to section 3.2.5 of the AMQP specification.

The Event Hubs documentation for `EventData` has been
updated to reflect this change.  Due to a known bug
in the Service Bus service, the service will reject
messages with `byte[]` application properties.  Until
this is confirmed fixed, the Service Bus documentation
still reflects `byte[]` as an invalid type for the
application properties.