[Event Hubs] ETW fixes
The focus of these changes is to fix several mistakes
related to ETW logs.  Some messages that included links
to the troubleshooting guide were improperly annotated
with `[Event]` instead of `[NonEvent]`, causing ETW
to log a warning for "too many parameters" when writing
the messages.  Also corrected a parameter ordering issue
for an event batch processing error, causing the structured
argument parameters to be ordered differently than they
appear in the message template.