ResponseBodyPolicy: ensure response content stream is MemoryStream if buffering is enabled
Provides a stronger guarantee about the behaviour of the response
content stream. Also make the requirement for unit tests stricter to
mimic the actual behaviour of the response stream, and disable async
stream checking in MockTransport for the case where we're guaranteed to
have a MemoryStream for the buffered content (which should allow
synchronous processing).