[Event Hubs] Update value of missing offset (#43659)
* [Event Hubs] Update value of missing offset

The focus of these changes is to update the value
written for a checkpoint when the offset is intentionally
not populated.  Previously, the offset was written as
empty, which was read by consumers as a null value.
With this change, the offset value is now written as
a string indicating that the offset is not present.

Because the checkpoint format changed and offset is
no longer populated by the EventProcessor, we need
to ensure that a value is present for the Functions
scale controller which uses a null check on the
offset to determine if a checkpoint is in the legacy
format or current.  Because GetCheckpointAsync will
only populate the offset if a long.TryParse is
successful, adding a nonsense string value to satisfy
the null check will not impact the EventProcessor behavior.