[CODEOWNERS] Removing retired library PR config (#43834)
The focus of these changes is to remove the PR configuration for the retired OpenTelemetry.LiveMetrics library.   Because the label still exists, the support entry has been left for now.

I'll discuss with the monitor team and see if there's a new label that should be mapped to existing/closed issues, otherwise, we will need to retain the label indefinitely for support/triage use.