Fix AzurePipelinesCredential in release pipelines
This is initially a bug that I made in my team's implementation of this
credential, which was used as the basis for this credential. If you
hard-code the hub as "build", the credential fails when used in a
release pipeline. This is just down-streaming the fix made on my team's
end.