[Messaging] Relax emulator endpoint restrictions (#44052)
The focus of these changes is to remove the restriction
that the endpoint used in connection strings must
resolve to a variant of `localhost` when using the
development emulator.

A fix to parsing was also made to be resilient to
an endpoint with no scheme that specifies a custom port.
Previously, this would parse incorrectly and be
rejected.  The scenario is now detected and parsed.

Also included is a package bump for the AMQP transport
library and release prep for the Event Hubs core package.