[Event Hubs] Diagnostics and validation tweaks
The focus of these changes is to fix a couple of
formatting issues in the ETW events and to enhance
the information provided when links are opened and
when the processor begins processing a partition.

In addition, startup validation logic has been
adjusted to use a custom set of options when testing
read permissions to avoid the default prefetch count
being used and transfer of more data than is needed
from the service.