[CODEOWNERS] Remove "all API changes" rule
The focus of these changes is to remove the rule that required the .NET architects to sign-off on all pull requests that resulted in a public API change.  

This rule predates the use of API View as the gate for API approvals and is redundant.  With the enforcement that a CODEOWNER for each area approve pull requests, this rule blocked them from closing out regardless of whether or not the associated API View had been approved.