[Functions] Prevent connection name mistake
The focus of these changes is to guard against a
common error made when configuring a trigger.  With
these changes, if the trigger's connection property
is set to a valid connection string instead of an
informational name, the mistake will be detected
and sensitive information will be redacted from the
error message to avoid accidental capture in logs
and similar mechanisms.