TransferManager fixes (#45714)
* channel reader tasks not static

channel readers tasks were static fields but held nonstatic method
execution and cancelled with object disposal.

* remove channel getters and setters in transfermanager

setters imply these are meant to change over object lifetime, but there
is no code that does so and such a change is incompatible with the
object lifetime tasks

* remove unused checkpointer options field

this was maybe leftover from when options were passed around for
reconstruction of checkpointer across packages? that doesn't seem to be
the case anymore. it's a private field and is never accessed outside of
constructor.

* unused property