Azure.Provisioning updates (#45786)
This update adds an overhaul to the earlier draft of our Azure.Provisioning libraries. Please note the PR is staged in two commits. I've put what I think are the most interesting files to review in the first commit and dumped all the generated code in the second. I'll also run through and add a quick guided tour of the interesting moving pieces since this is a lot to take in all at once.