[Event Hubs] Resilient management link creation
The focus of these changes is to correctly handle the scenario where the
management AMQP link is in the process of closing (such as for an idle timeout)
as a new operation attempts to use it.  This error should be detected as a special
case and retried as it is with producer and consumer operations.