[Event Hubs] GeoDR - OffsetString (#47133)
* [Event Hubs] GeoDR - OffsetString

The focus of these changes is to introduce members for the string-based offset
format and mark the numeric-based offset members as obsolete.  This is part of
the Event Hubs GeoDR feature design for the service, where replication-based
offsets will be composite strings and no longer parsable to numeric values.

Not included in these changes is the error handling for the new error for an
invalid reader position after failover.  The error details are still being
finalized by the service and will be included in a future change set.

* Fixing AOT warning

* Fixing build error from rebase and regenerating net8.0 API listings.

* Fixing comments

---------

Co-authored-by: Madalyn Redding <66138537+m-redding@users.noreply.github.com>