Create minimum buffer required for OpenRead
LazyLoadingReadOnlyStream, when not specified an explicit buffer size,
always selects 4MB as the buffer size. This is unnecessary if the
underlying blob size (which comes in as initialLength) is smaller than
that AND the blob does not allow modifications (meaning the blob will
never grow any larger). In this case, the buffer size can simply be
initialized to the length of the blob that will be read.