Implement .NET AI interfaces on ChatCompletionsClient / EmbeddingsClient
This moves over the implementations currently in Microsoft.Extensions.AI.AzureAIInference. That package has been temporary; the intent is these Azure.AI.Inference types will be directly usable without requiring or starting with another package, and the implementations can be kept up-to-date with Azure.AI.Inference as it evolves. These implementations enable Azure.AI.Inference to be used by all clients that expect an IChatClient / IEmbeddingGenerator, plus with all middleware that works in terms of these interfaces, extending the reach and applicability of the library.