[Service Bus] Fix error parsing logic
The focus of these changes is to fix the logic for
parsing error codes from the Service Bus administration
service.  The implementation assumes that all error
responses will have a content body, which is not always
the case.

For example, an HTTP 401 is returned without a body.
In this scenario, the error detail parsing faults
and surfaces an exception, obscuring the original
401 status code.