[SignalR] Fixed transitive version conflict
The focus of these changes is to fix a transitive
version conflict in the SignalR extension, arising
from the legacy `Microsoft.Azure.Functions.Extensions`
package and its dependency on an old version of
`Microsoft.Extensions.DependencyInjection" which
conflicts with the 8.x package line.

This was added as a local override, as SignalR is
the only Functions extension package in the repository
which uses the legacy package.