[Event Hubs] Set GeoDR on management link
The focus of these changes is to set the desired capability for GeoDR on the
management link, so that requests for Event Hub and partition metadata can take
it into account for offset formatting.  Also included are tweaks to the tests for
`EventPosition.Latest` to reduce flakininess due to timing issues during nightly
runs.