[Hotfix: Event Hubs] Processor release prep (#48264)
* [Event Hubs] Fixed producer semaphore release (#47965)

* [Event Hubs] Fixed producer semaphore release

The focus of these changes is to fox an obscure edge case in the
`EventHubBufferedProducer` client where an obscure race condition when
flushing/enqueuing events concurrently with disposing the producer could
cause a semaphore to be released inappropriately.  This error superseded
the `TaskCanceledException` that should have been surfaced.

Also moved to volatile reads for the total buffer count when making
state-related decisions, to ensure that the most recent value is used.