[Extensions] Fix ASP.NET Core integration testing
The focus of these changes is to fix an assumption
in the logic when creating clients from configuration.
Currently, if the configuration section has a non-null
value, it is assumed to be a connection string.
This is not the case for ASP.NET Core test integration
due to a bug in the test host which causes the Value
property to return an empty string rather than the
null returned by the actual configuration system.