[Event Hubs] Fix trigger data label (#48517)
The focus of these changes is to fix a misspelling of "PartitionId" in the
trigger input data passed to the function executor.  This caused function logs
and metrics reported by AppInsights and the portal to reflect the wrong label.
To ensure that applications that rely on the misspelling are not impacted, a new
member with the correct spelling was added.