[Messaging] Minor updates for April 2025 release
The focus of these changes is to make a small set
of minor updates to the Azure Messaging libraries.
Included are:

- Clarifying the load balancing behavior of the
Event Hubs processor's load balancing strategies.

- Enhancing the retry policies for better support
in web socket scenarios.  It was observed that the
relevant failure information in many scenarios was
embedded in an inner `WebSocketException` and
wrapped by `WebSocketException`.  Going forward,
retry decisions will be made based on the inner
exception, if present.

- Added an option for a remote certificate validation
callback to the `ServiceBusClientOptions` in order
to allow applications to control validation behavior
and to align with the existing option in Event Hubs.