[Event Hubs] Preserve system property broker types
The focus of these changes is to fix a bug where
the data types of broker-owned properties were
being adjusted when an event was read by the
client, causing the underlying AMQP data to be
mutated.

This resulted in binary changes when the AMQP
message was serialized and unintentionally altered
the service contract. Going forward, the original
original data types will be preserved on the AMQP
representation of the message and type
normalization only applied to the .NET EventData
projection.