[Extensions] Package updates and test fixes
The focus of these changes is to complete updating
the extensions package reference versions and to
address a breaking change in the
Microsoft.Extensions.Logging.Abstractions package.
Because the NullLogger class used by the Azure
Functions test host was moved to the
Microsoft.Extensions.Logging package, test projects
need to add the non-abstraction package reference
directly until the host is updated.