[Event Hubs Trigger] Work around service contract (#50645)
* [Event Hubs Trigger] Work around service contract

The focus of these changes is to work around the
service contract for `offset` which is a string
that previously was always numeric.  In support
of the GeoDR feature, the service no longer
returns a numeric offset in all scenarios, sometimes
returning a format that cannot be parsed as a number.

To avoid applications that made assumptions about
the format, the `Offset` binding contracts for the
trigger were updated to always return a numeric
value as the string.  When the service returns an
offset that cannot be parsed as a number, the value
`-1` is substituted.  If used for readers, this would
possibly result in reading at the beginning of the
stream, causing duplicate processing of events but
will ensure there is no data loss.  Because the
`Offset` binding is not longer guaranteed to represent
the actual offset, a new `OffsetString` binding was
introduced to surface the actual offset value.