Relax JSON encoding in System.ClientModel (#50885)
The default encoding rules used by Utf8JsonWriter are very strict, more than what's needed just for JSON, under the assumption that the resulting JSON might be embedded in an HTML page / script element. That strictness is not necessary or desirable when JSON is just being sent as part of an application/json payload, as it is with API inputs/outputs. The stricter encoding rules can have measurable impact, especially with the kinds of payloads commonly used in these APIs, e.g. base64 includes '+' in its alphabet, and '+' is a character that's encoded in the strict mode and not in the relaxed mode.

This PR relaxes the encoding rules used.