[Ingestion] AOT Support
The focus of these changes is to regenerate the
Ingestion client using the new AOT-friendly serialization
paths and to ensure that the client is compatible
with Ahead-of-Time (AOT) compilation.  This is
critical for inclusion in the Azure.MCP product.

Compliance required new overloads for log uploads,
as the existing form allowed for runtime serialization
in different forms, which were not compatible with AOT.
For AOT scenarios, log uploads will accept `BinaryData`
directly, allowing callers to prepare their data
in a way that is compatible with AOT compilation.