[EngSys] Update Analyzers version (#51232)
* [EngSys] Update Analyzers version

The focus of these changes is to update the Azure
analyzers to the latest version, which refines
messages and naming suggestions.

Included are AZC0034 (Duplicate type names) and
AZC0035 (Missing model factory methods) which had
been previously added but were never released.  A
number of temporary suppressions have been added
to allow the analyzers to be present without causing
build failures.  GitHub issues have been created to
track the work of fixing the underlying causes.

* Exporting API to fix pipeline failures

* More fixes