Fix incorrectly merged models
Some models were generated with the same C# class,
causing a race condition when the generator tried to
write the same file twice.

This uses `remove-model` to remove the old versions
of these models. Some generated code has been moved
into custom classes to ensure we keep the mapping of
the old and new model name to the same C# type on
the deserialization path.

Uses `CodeGenSerialization` attribute to include
generated deserialization code for bufferSize, which
was defined by the old model.