Remove invalid error check from PersistentAgentsChatClient
The check is overly strict; it assumes all of the provided messages are in response to a previous conversation with this service, but the messages could be the history from a conversation with another service, as can happen in agent-based scenarios. In such cases, there may be function tool calls and results in the history; the rest of the implementation will properly ignore them, as the service doesn't provide any way to send those calls into the service, but this check was blowing up before then.