Fix duplication between agents pre-configured tools and ChatOptions.Tools (#53041)
Azure.AI.Agents.Persistent lets you configure an agent with function signatures it's implicitly aware of, rather than them needing to be provided per request. That, however, creates complication for callers via IChatClient, as if they provide that function in ChatTools.Options, it leads to the function's signature being sent as part of the request, and the duplication of it with the pre-configured function signature results in an error. It's possible to work around this, by simply not including that function in the request, but it's a natural thing for a developer to do, especially if they want the function to be automatically invoked when the model requests it. You can achieve that by putting the function into the FunctionInvocationChatClient's AdditionalTools, which exists for this kind of purpose, but that's harder to discover.

Rather than try to do something complicated, this commit just deduplicates all function tools such that rather than getting an error when multiple tools have the same name, only one of them is kept.

Co-authored-by: Roger Barreto <19890735+rogerbarreto@users.noreply.github.com>