[ClientModel] Fix RetryPolicy delay calculations
The focus of these changes is to correct a mismatch
between the "try count" parameter used for declay
calculations, which assumes that there has been at
least the initial attempt but is being passed the
current retry count which which is zero-based.  This
resulted in the delay for the first retry being
negative and the retry taking place immediately.